refactor: tighten cli boundaries and delete hygiene
