+++ src/lifeos_cli/db/services/__init__.py
@@ -1,142 +1,1 @@
-"""Database-backed services for lifeos_cli."""
-
-from .areas import (
-    AreaAlreadyExistsError,
-    AreaNotFoundError,
-    batch_delete_areas,
-    create_area,
-    delete_area,
-    get_area,
-    list_areas,
-    update_area,
-)
-from .batching import BatchDeleteResult
-from .entity_tags import load_tags_for_entities, sync_entity_tags
-from .notes import (
-    NoteBatchDeleteResult,
-    NoteBatchUpdateResult,
-    NoteNotFoundError,
-    batch_delete_notes,
-    batch_update_note_content,
-    create_note,
-    delete_note,
-    get_note,
-    list_notes,
-    search_notes,
-    update_note,
-)
-from .people import (
-    PersonAlreadyExistsError,
-    PersonNotFoundError,
-    batch_delete_people,
-    create_person,
-    delete_person,
-    get_person,
-    list_people,
-    update_person,
-)
-from .tags import (
-    VALID_TAG_ENTITY_TYPES,
-    InvalidTagEntityTypeError,
-    TagAlreadyExistsError,
-    TagNotFoundError,
-    batch_delete_tags,
-    create_tag,
-    delete_tag,
-    get_tag,
-    list_tags,
-    normalize_tag_name,
-    update_tag,
-    validate_tag_entity_type,
-)
-from .tasks import (
-    InvalidPlanningCycleError,
-    InvalidTaskDepthError,
-    ParentTaskReferenceNotFoundError,
-    TaskNotFoundError,
-    VisionReferenceNotFoundError,
-    batch_delete_tasks,
-    create_task,
-    delete_task,
-    get_task,
-    list_tasks,
-    update_task,
-)
-from .visions import (
-    AreaReferenceNotFoundError,
-    VisionAlreadyExistsError,
-    VisionNotFoundError,
-    batch_delete_visions,
-    create_vision,
-    delete_vision,
-    get_vision,
-    list_visions,
-    update_vision,
-    validate_vision_status,
-)
-
-__all__ = [
-    "AreaAlreadyExistsError",
-    "AreaNotFoundError",
-    "AreaReferenceNotFoundError",
-    "BatchDeleteResult",
-    "InvalidPlanningCycleError",
-    "InvalidTagEntityTypeError",
-    "InvalidTaskDepthError",
-    "NoteBatchDeleteResult",
-    "NoteBatchUpdateResult",
-    "NoteNotFoundError",
-    "ParentTaskReferenceNotFoundError",
-    "PersonAlreadyExistsError",
-    "PersonNotFoundError",
-    "TagAlreadyExistsError",
-    "TagNotFoundError",
-    "batch_delete_areas",
-    "batch_delete_people",
-    "batch_delete_notes",
-    "batch_delete_tags",
-    "batch_delete_tasks",
-    "batch_delete_visions",
-    "batch_update_note_content",
-    "create_area",
-    "create_note",
-    "create_person",
-    "create_tag",
-    "create_task",
-    "create_vision",
-    "delete_area",
-    "delete_note",
-    "delete_person",
-    "delete_tag",
-    "delete_task",
-    "delete_vision",
-    "get_note",
-    "get_area",
-    "get_person",
-    "get_tag",
-    "get_task",
-    "get_vision",
-    "list_notes",
-    "list_areas",
-    "list_people",
-    "list_tags",
-    "list_tasks",
-    "list_visions",
-    "load_tags_for_entities",
-    "normalize_tag_name",
-    "search_notes",
-    "sync_entity_tags",
-    "TaskNotFoundError",
-    "VALID_TAG_ENTITY_TYPES",
-    "validate_tag_entity_type",
-    "validate_vision_status",
-    "VisionAlreadyExistsError",
-    "VisionNotFoundError",
-    "VisionReferenceNotFoundError",
-    "update_area",
-    "update_note",
-    "update_person",
-    "update_tag",
-    "update_task",
-    "update_vision",
-]
+"""Database service modules for lifeos_cli."""
