+++ tests/test_cli_domains.py
@@ -8,13 +8,8 @@ from uuid import UUID
 import pytest
 
 from lifeos_cli import cli
-from lifeos_cli.cli_support import (
-    area_handlers,
-    people_handlers,
-    tag_handlers,
-    task_handlers,
-    vision_handlers,
-)
+from lifeos_cli.db import session as db_session
+from lifeos_cli.db.services import areas, people, tags, tasks, visions
 
 
 def test_main_area_add_creates_area(
@@ -29,8 +24,8 @@ def test_main_area_add_creates_area(
         assert kwargs["name"] == "Health"
         return SimpleNamespace(id=UUID("11111111-1111-1111-1111-111111111111"))
 
-    monkeypatch.setattr(area_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(area_handlers, "create_area", fake_create_area)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(areas, "create_area", fake_create_area)
 
     exit_code = cli.main(["area", "add", "Health"])
     captured = capsys.readouterr()
@@ -52,8 +47,8 @@ def test_main_tag_add_creates_tag(
         assert kwargs["entity_type"] == "person"
         return SimpleNamespace(id=UUID("22222222-2222-2222-2222-222222222222"))
 
-    monkeypatch.setattr(tag_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(tag_handlers, "create_tag", fake_create_tag)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(tags, "create_tag", fake_create_tag)
 
     exit_code = cli.main(["tag", "add", "family", "--entity-type", "person"])
     captured = capsys.readouterr()
@@ -89,8 +84,8 @@ def test_main_people_show_prints_tags(
             deleted_at=None,
         )
 
-    monkeypatch.setattr(people_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(people_handlers, "get_person", fake_get_person)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(people, "get_person", fake_get_person)
 
     exit_code = cli.main(["people", "show", "33333333-3333-3333-3333-333333333333"])
     captured = capsys.readouterr()
@@ -112,8 +107,8 @@ def test_main_vision_add_creates_vision(
         assert kwargs["name"] == "Launch lifeos-cli"
         return SimpleNamespace(id=UUID("44444444-4444-4444-4444-444444444444"))
 
-    monkeypatch.setattr(vision_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(vision_handlers, "create_vision", fake_create_vision)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(visions, "create_vision", fake_create_vision)
 
     exit_code = cli.main(["vision", "add", "Launch lifeos-cli"])
     captured = capsys.readouterr()
@@ -135,8 +130,8 @@ def test_main_task_add_creates_task(
         assert kwargs["priority"] == 2
         return SimpleNamespace(id=UUID("55555555-5555-5555-5555-555555555555"))
 
-    monkeypatch.setattr(task_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(task_handlers, "create_task", fake_create_task)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(tasks, "create_task", fake_create_task)
 
     exit_code = cli.main(
         [
@@ -178,8 +173,8 @@ def test_main_task_batch_delete_prints_summary(
             errors=(),
         )
 
-    monkeypatch.setattr(task_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(task_handlers, "batch_delete_tasks", fake_batch_delete_tasks)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(tasks, "batch_delete_tasks", fake_batch_delete_tasks)
 
     exit_code = cli.main(
         [
@@ -217,8 +212,8 @@ def test_main_people_batch_delete_reports_missing_ids(
             errors=("Person 33333333-3333-3333-3333-333333333333 was not found",),
         )
 
-    monkeypatch.setattr(people_handlers, "session_scope", fake_session_scope)
-    monkeypatch.setattr(people_handlers, "batch_delete_people", fake_batch_delete_people)
+    monkeypatch.setattr(db_session, "session_scope", fake_session_scope)
+    monkeypatch.setattr(people, "batch_delete_people", fake_batch_delete_people)
 
     exit_code = cli.main(
         [
